refactor(engine): fold orchestrator into engine as src/orchestrator
Slim-down phase 3. The orchestrator package's only consumers were the
engine's built-in agent-teams-plugin (3 import sites) and cli's /team
commands (1 file); it had zero specs (--passWithNoTests) and carried a
root tsconfig alias plus a workspace of its own for 695 LOC.

- move packages/orchestrator/src -> packages/engine/src/orchestrator
  (deps: zod only, already an engine dep; no reverse engine imports,
  so no cycle)
- engine: rewrite the 3 built-in import sites to relative paths; add
  tsup entry orchestrator/index + package.json subpath export
  "./orchestrator" (public API unchanged, dist/orchestrator/index.*)
- cli: team-commands.ts now imports pulse-coder-engine/orchestrator;
  drop the workspace dep
- remote-server: prebuild no longer builds the orchestrator filter
- root tsconfig: remove both pulse-coder-orchestrator aliases
- preserve module knowledge as engine harness/knowledge/orchestrator.md
  (positioning, AgentRunner boundary, DAG invariants, artifact-store
  path, zero-spec warning) + engine AGENTS.md nav row
- sync root AGENTS.md (§1 route, §2 aliases/tests, §5), READMEs (EN/CN),
  agent-teams contracts note, cli/engine READMEs, roadmap skills

Verified: engine typecheck + dts green; build green; test:core 471 green;
agent-teams typecheck (dist canary) green; canvas-workspace typecheck +
full suite green (2253 tests); check-harness harnessGaps: 0 (10/10).
