+++ apps/remote-server/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "type": "module",
   "scripts": {
-    "prebuild": "pnpm --filter pulse-coder-orchestrator build && pnpm --filter pulse-coder-engine build && SKIP_DTS=1 pnpm --filter pulse-coder-acp --filter pulse-coder-memory-plugin --filter pulse-coder-plugin-kit --filter pulse-coder-langfuse-plugin build",
+    "prebuild": "pnpm --filter pulse-coder-engine build && SKIP_DTS=1 pnpm --filter pulse-coder-acp --filter pulse-coder-memory-plugin --filter pulse-coder-plugin-kit --filter pulse-coder-langfuse-plugin build",
     "dev": "tsx watch src/index.ts",
     "build": "tsup",
     "pretest": "SKIP_DTS=1 pnpm --filter pulse-coder-plugin-kit build",
