+++ packages/cli/package.json
@@ -24,7 +24,6 @@
     "pulse-coder-agent-teams": "workspace:*",
     "pulse-coder-engine": "workspace:*",
     "pulse-coder-memory-plugin": "workspace:*",
-    "pulse-coder-orchestrator": "workspace:*",
     "react": "^19.2.5",
     "zod": "^4.3.6"
   },
