+++ packages/engine/tsup.config.ts
@@ -3,7 +3,8 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
-    'built-in/index': 'src/built-in/index.ts'
+    'built-in/index': 'src/built-in/index.ts',
+    'orchestrator/index': 'src/orchestrator/index.ts'
   },
   format: ['esm', 'cjs'],
   dts: true,
