--- packages/orchestrator/harness/validate/validation.yaml
@@ -1,12 +1,0 @@
-version: 1
-
-pathRules:
-  - name: default
-    paths:
-      - src/**
-      - package.json
-      - tsconfig.json
-      - tsup.config.ts
-    required:
-      - pnpm --filter pulse-coder-orchestrator test
-      - pnpm --filter pulse-coder-orchestrator typecheck
