--- packages/orchestrator/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: { index: 'src/index.ts' },
-  format: ['esm', 'cjs'],
-  dts: true,
-  clean: true,
-  splitting: false,
-  sourcemap: true,
-  target: 'es2022'
-});
